Using mapDispatchToProps() Shorthand Notation.
Avoiding the boilerplate code in mapDispatchToProps() for the common case where action creator arguments match the callback prop arguments

When the arguments for the callback prop match the arguments to the action creator exactly, there is a shorter way to specify mapDispatchToProps.
Rather than pass a function, we can pass an object mapping of the names of the callback props that we want to inject and the action creator functions that create the corresponding actions.
This is a rather common case, so often you don't need to write mapDispatchToProps, and you can pass this map in object instead.